search: move internal structural search logic to own file
